UGE_SIA/mas_trading_llm/requirements.txt
numpy
pandas
python-dotenv
google-genai
